[auth-swift] AuthStorage protocol in own file; split AuthKeychainServices into two types
