+++ FirebaseAuth/Sources/Swift/Storage/AuthKeychainServices.swift
@@ -205,80 +205,4 @@ private let kAccountPrefix = "firebase_auth_1_"
       kSecAttrAccount as String: key,
     ]
   }
-
-  // MARK: - Private methods for shared keychain operations
-
-  // TODO(ncooke3): The following APIs are essentially duplicates of the API
-  // defined by the `AuthStorage` protocol. It'd be nice to consolidate the
-  // two sets into a single API surface. They were added in
-  // https://github.com/firebase/firebase-ios-sdk/pull/2684
-
-  /** @fn getItemWithQuery:error:
-   @brief Get the item from keychain by given query.
-   @param query The query to query the keychain.
-   @return The item of the given query. nil if not exsit.
-   */
-  @objc public func getItem(query: [String: Any]) throws -> DataWrapper {
-    var mutableQuery = query
-    mutableQuery[kSecReturnData as String] = true
-    mutableQuery[kSecReturnAttributes as String] = true
-    mutableQuery[kSecMatchLimit as String] = 2
-
-    var result: AnyObject?
-    let status =
-      SecItemCopyMatching(mutableQuery as CFDictionary, &result)
-
-    if let items = result as? [[String: Any]], status == noErr {
-      if items.count != 1 {
-        throw AuthErrorUtils.keychainError(function: "SecItemCopyMatching", status: status)
-      }
-
-      return DataWrapper(data: items[0][kSecValueData as String] as? Data)
-    }
-
-    if status == errSecItemNotFound {
-      return DataWrapper(data: nil)
-    } else {
-      throw AuthErrorUtils.keychainError(function: "SecItemCopyMatching", status: status)
-    }
-  }
-
-  /** @fn setItem:withQuery:error:
-   @brief Set the item into keychain with given query.
-   @param item The item to be added into keychain.
-   @param query The query to query the keychain.
-   @return Whether the operation succeed.
-   */
-  @objc public func setItem(_ item: Data, withQuery query: [String: Any]) throws {
-    let status: OSStatus
-    let function: String
-    if (try getItem(query: query)).data != nil {
-      let attributes: [String: Any] = [kSecValueData as String: item]
-      status = SecItemUpdate(query as CFDictionary, attributes as CFDictionary)
-      function = "SecItemUpdate"
-    } else {
-      var queryWithItem = query
-      queryWithItem[kSecValueData as String] = item
-      status = SecItemAdd(queryWithItem as CFDictionary, nil)
-      function = "SecItemAdd"
-    }
-
-    if status == noErr {
-      return
-    }
-    throw AuthErrorUtils.keychainError(function: function, status: status)
-  }
-
-  /** @fn getItemWithQuery:error:
-   @brief Remove the item with given queryfrom keychain.
-   @param query The query to query the keychain.
-   @return Whether the operation succeed.
-   */
-  @objc public func removeItem(query: [String: Any]) throws {
-    let status = SecItemDelete(query as CFDictionary)
-    if status == noErr || status == errSecItemNotFound {
-      return
-    }
-    throw AuthErrorUtils.keychainError(function: "SecItemDelete", status: status)
-  }
 }
