+++ FirebaseAuth/Sources/Swift/Storage/AuthStorage.swift
@@ -1,0 +1,36 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import Foundation
+
+/// A protocol for a persistent data storage container.
+@objc(FIRAuthStorage) public protocol AuthStorage: NSObjectProtocol {
+  /// Retrieves the data for the given key from storage.
+  /// - Parameter key: The key to use.
+  /// - Returns: The data stored in the storage for the given key.
+  /// - Throws: An error if the operation is not successful.
+  @objc func data(forKey key: String) throws -> DataWrapper
+
+  /// Sets the data for the given key from storage.
+  /// - Parameters:
+  ///   - data: The data to store.
+  ///   - key: The key to use.
+  /// - Throws: An error if the operation is not successful.
+  @objc func setData(_ data: Data, forKey key: String) throws
+
+  /// Removes the data for the given key from storage.
+  /// - Parameter key: The key to use.
+  /// - Throws: An error if the operation is not successful.
+  @objc func removeData(forKey key: String) throws
+}
