+++ FirebaseAuth/Sources/Swift/Storage/AuthUserDefaults.swift
@@ -27,48 +27,6 @@ private let kPersistentDomainNamePrefix = "com.google.Firebase.Auth."
   }
 }
 
-/**
-    @brief The protocol for permanant data storage.
- */
-@objc(FIRAuthStorage) public protocol AuthStorage: NSObjectProtocol {
-  // NOTE: Removing this from protocol since it's not being used anywhere.
-  /** @fn initWithService:
-      @brief Initialize a @c FIRAuthStorage instance.
-      @param service The name of the storage service to use.
-      @return An initialized @c FIRAuthStorage instance for the specified service.
-   */
-  // - (id<FIRAuthStorage>)initWithService:(NSString *)service;
-
-  /** @fn dataForKey:error:
-      @brief Gets the data for @c key in the storage. The key is set for the attribute
-          @c kSecAttrAccount of a generic password query.
-      @param key The key to use.
-      @param error The address to store any error that occurs during the process, if not NULL.
-          If the operation was successful, its content is set to @c nil .
-      @return The data stored in the storage for @c key, if any.
-   */
-  @objc func data(forKey key: String) throws -> DataWrapper
-
-  /** @fn setData:forKey:error:
-      @brief Sets the data for @c key in the storage. The key is set for the attribute
-          @c kSecAttrAccount of a generic password query.
-      @param data The data to store.
-      @param key The key to use.
-      @param error The address to store any error that occurs during the process, if not NULL.
-      @return Whether the operation succeeded or not.
-   */
-  @objc func setData(_ data: Data, forKey key: String) throws
-
-  /** @fn removeDataForKey:error:
-      @brief Removes the data for @c key in the storage. The key is set for the attribute
-          @c kSecAttrAccount of a generic password query.
-      @param key The key to use.
-      @param error The address to store any error that occurs during the process, if not NULL.
-      @return Whether the operation succeeded or not.
-   */
-  @objc func removeData(forKey key: String) throws
-}
-
 /** @class FIRAuthUserDefaults
     @brief The utility class to storage data in NSUserDefaults.
  */
